Scan complete message is now displayed as soon as a record is saved.
Also I replaced the countdown with a progress bar - better suited for
a non-linear progress.
new message type 'puck scan completed' was added - this is mainly fo me
to know what is the source of the message.
Main_manager is very complicated. It has too many 'ifs' :/
A bigger rewrite is needed for the class soon.